requirements.txt
diffusers
fastapi
Pillow
torch
transformers
accelerate
uvicorn
autoawq
